added stock finder component
