+++ src/constants/api.ts
@@ -1,0 +1,2 @@
+export const proxyServerUrl =
+  process.env.REACT_APP_PROXY_SERVER || 'https://cors-proxy.fringe.zone/'
